Updated for 1.4.7
Updated for 1.4.7, and probably every other recent version (untested).
Removed lighting support. This had broken dependencies on CraftBukkit
that I'm not smart enough to fix. It may be possible to implement code
specific to each release of the MC server to achieve the same effect,
but I'm not sure that's a good idea. We'll see.
Requesting lighting in commands should silently fail, so any scripts
should work as intended.
This release is largely untested. As such, I've gone ahead and changed
the contact info for the plugin to prevent Nephyrin from getting angry
emails about crashes, etc.